Moved codegen tests to separate module
Nightly and stable produce different outputs, thus only test nightly.
