+++ integration_tests/codegen_fail/fail/enum/derive_no_fields.rs
