+++ integration_tests/codegen_fail/fail/input-object/derive_incompatible_object.stderr
