+++ integration_tests/codegen_fail/fail/input-object/derive_no_underscore.stderr
