+++ integration_tests/codegen_fail/fail/input-object/derive_unique_name.rs
