+++ integration_tests/codegen_fail/fail/interface/impl_argument_no_object.rs
