+++ integration_tests/codegen_fail/fail/interface/impl_no_fields.rs.disabled
