+++ integration_tests/codegen_fail/fail/interface/impl_no_input_object.rs.disabled
