+++ integration_tests/codegen_fail/fail/interface/impl_unqiue_name.rs.disabled
