+++ integration_tests/codegen_fail/fail/object/derive_fields_unique.stderr
