+++ integration_tests/codegen_fail/fail/object/derive_incompatible_input_object.rs
