+++ integration_tests/codegen_fail/fail/object/derive_no_fields.rs
