+++ integration_tests/codegen_fail/fail/object/impl_argument_no_object.rs
