+++ integration_tests/codegen_fail/fail/object/impl_fields_unique.stderr
