+++ integration_tests/codegen_fail/fail/object/impl_incompatible_input_object.rs
