+++ integration_tests/codegen_fail/fail/object/impl_no_argument_underscore.stderr
