+++ integration_tests/codegen_fail/fail/object/impl_no_fields.stderr
