+++ integration_tests/codegen_fail/fail/union/derive_enum_field.rs
