+++ integration_tests/codegen_fail/fail/union/derive_no_fields.rs
