+++ integration_tests/codegen_fail/fail/union/derive_same_type.rs
