+++ integration_tests/codegen_fail/fail/union/impl_enum_field.rs
