+++ integration_tests/codegen_fail/fail/union/impl_no_fields.rs
