+++ integration_tests/codegen_fail/fail/union/impl_same_type.rs
